test_data_hub_backend-main/requirements.txt
openai>=1.12.0
pandas>=2.0.0,<2.2.3
python-dotenv==1.0.0
jsonschema==4.19.0
faker==19.6.2
numpy>=1.25.0,<2.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
cx-Oracle==8.3.0
requests>=2.31.0
